Move pick method to common as we can use it in the app as well
