+++ common/utils/pick.js
@@ -1,10 +1,8 @@
 export default function pick(fields, obj) {
-  const res = fields.reduce((result, fieldName) => {
+  return fields.reduce((result, fieldName) => {
     if (typeof obj[fieldName] !== 'undefined') {
       result[fieldName] = obj[fieldName]
     }
     return result
   }, {})
-
-  return res
 }
