+++ server/controllers/auth.js
@@ -4,7 +4,7 @@ import { Op } from 'sequelize'
 import User from '../models/user'
 import checkout from '../services/checkout'
 import db from '../database'
-import pick from '../utils/pick'
+import pick from '../../common/utils/pick'
 import { APIError } from '../helpers/errors'
 import { executePayment } from '../services/paypal'
 import { generateRandomHash } from '../utils/randomHash'
