+++ server/controllers/event.js
@@ -17,7 +17,7 @@ import {
   addRequestResourcesActivity,
 } from '../services/activity'
 
-import pick from '../utils/pick'
+import pick from '../../common/utils/pick'
 import { APIError } from '../helpers/errors'
 import { createEventSlots, isInClosedOrder } from '../../common/utils/slots'
 import { getConfig } from '../config'
